\name{Ginternum}
\alias{Ginternum}
\title{plot the number of interactions}
\description{
  Take the data and plot the number of sinlg, pair, triple and
  quadruple interactions.
}
\usage{
Ginternum(data)
}
\arguments{
  \item{data}{cCGH data}
}
\examples{
Ginternum(data)
}
\value{
  plot the number of interactions
}
\keyword{plots}
