\name{Gnet}
\alias{Gnet}
\title{Simulate a network}
\description{
  From candidate sets construct a network having less or equal to tmaxe
  ensembles.
}
\usage{
Gnet(candid,tmaxe)
}
\arguments{
  \item{candid}{candidate sets from Gcandi}
  \item{tmaxe}{maximum number of ensemble in a network}
}
\value{
    general network
}
