\name{rcc}
\alias{rcc}
\non_function{}
\title{Renal Cell Carcinomas (RCC) cCGH data}
\usage{data(rcc)}
\format{
  RCC data was collected at 82 aberrations from 116 tumors by Jiang and Moch (Jiang {\em et al.}, 2000).
}
\description{
  Chromosome loss or gain was measured in each chromosome
  arm. Aberration is indicated as 1 otherwise 0.
}
\source{Jiang, F., Desper, R., Papadimitriou, C.H., Schaffer, A.A., Kallioniemi, O.P., Richter, J., Schraml, P., Sauter, G., Mihatsch, M.J., Moch, H.(2000), Construction of evolutionary tree models for renal cell carcinoma from comparative genomic hybridization data, Cancer Research, 60, 6503-6509.  
}
\examples{
data(rcc)
}
\keyword{datasets}
